# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Tutor API",
    "description": "API for AI Tutor, managing students, courses, and AI-generated questions. For more details, visit the [GitHub repository](https://github.com/ShishirShekhar/E-Learning-Platform.git).",
    "version": "1.0.0",
    "contact": {
      "name": "Shishir Shekhar",
      "email": "sspdav02@gmail.com"
    },
    "termOfService": "https://www.google.com",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local server"
    }
  ],
...
  "paths": {
...
    "/api/v1/user/me": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get user",
        "description": "Endpoint to get user",
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "string"
                        },
                        "email": {
                          "type": "string",
                          "example": "string"
                        },
                        "role": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "string"
                        },
                        "email": {
                          "type": "string",
                          "example": "string"
                        },
                        "role": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "error": {}
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "User not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "User not found"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {},
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}